Commit (11)
Added in an options setting for the console. moved 'Edit Company
details' there, added a change background and foreground colour feature
there too. set up the save and load feature for the colour so you dont
have to set it each time (Phew, that would be a pain in the arse).
Had some issues with it loading the colours at the start, so all I had
to do was move the LoadOptions() sub to be the first thing it does)